[CHANGE] Criação da API

+++ src/SoftApplication.IO.Domain/Eventos/Commands/RegistrarEventoCommand.cs
@@ -2,7 +2,7 @@
 {
     public class RegistrarEventoCommand : BaseEventoCommand
     {
-        public RegistrarEventoCommand(string pNome, double pValorCalculado)
+        public RegistrarEventoCommand(string pNome, decimal pValorCalculado)
         {
             Nome = pNome;
             ValorCalculado = ValorCalculado;
